# movietown/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a mistake.",
        "title": "MovieTown API",
        "contact": {},
        "version": "0.1-alpha"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/movies": {
            "get": {
                "description": "get []model.Movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Show movies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Movie"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/movies/{movie_id}": {
            "get": {
                "description": "get []model.Movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Show movie info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "movie_id",
                        "name": "movie_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Movie"
                        }
                    }
                }
            }
        },
        "/api/v1/screening": {
            "get": {
                "description": "get []Screening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screening"
                ],
                "summary": "Show screenings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "first",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "second",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Screening"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/screening/add": {
            "post": {
                "description": "Boss is able to add a new screening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screening"
                ],
                "summary": "Add new screening",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "add screening DTO",
                        "name": "screening",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.newScreening"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/screening/{movie_id}": {
            "get": {
                "description": "get []Screening by movie_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "screening"
                ],
                "summary": "Show screenings",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "screening search by movie_id",
                        "name": "movie_id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "first",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "second",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Screening"
                            }
                        }
                    }
                }
            }
        },
        "/v1/customer/reservations": {
            "get": {
                "description": "get []Reservation by customer_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Show reservations",
                "parameters": [
                    {
                        "type": "string",
                        "format": "int",
                        "description": "reservation search by customer_id",
                        "name": "customer_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reservation"
                        }
                    }
                }
            }
        },
        "/v1/customer/reservations/{customer_id}": {
            "get": {
                "description": "get []Reservation by customer_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Show reservations",
                "parameters": [
                    {
                        "type": "string",
                        "format": "int",
                        "description": "reservation search by customer_id",
                        "name": "customer_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reservation"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.newScreening": {
            "type": "object",
            "required": [
                "movie_hall_id",
                "movie_id",
                "start_of_screening"
            ],
            "properties": {
                "movie_hall_id": {
                    "type": "integer"
                },
                "movie_id": {
                    "type": "integer"
                },
                "start_of_screening": {
                    "type": "string"
                }
            }
        },
...
        "model.Customer": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_student": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.Employee": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/model.Role"
                },
                "roleId": {
                    "type": "integer"
                },
                "surname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "model.Movie": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "director": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "length": {
                    "type": "integer"
                },
                "movieType": {
                    "$ref": "#/definitions/model.MovieType"
                },
                "movieTypeId": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.MovieHall": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number_of_seats": {
                    "type": "integer"
                }
            }
        },
        "model.MovieType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.Reservation": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "customer": {
                    "$ref": "#/definitions/model.Customer"
                },
                "customerId": {
                    "type": "integer"
                },
                "employee": {
                    "$ref": "#/definitions/model.Employee"
                },
                "employeeId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "paid": {
                    "type": "boolean"
                },
                "reservationType": {
                    "$ref": "#/definitions/model.ReservationType"
                },
                "reservationTypeId": {
                    "type": "integer"
                },
                "reserved": {
                    "type": "boolean"
                },
                "screening": {
                    "$ref": "#/definitions/model.Screening"
                },
                "screeningId": {
                    "type": "integer"
                }
            }
        },
        "model.ReservationType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "model.Screening": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "movie": {
                    "$ref": "#/definitions/model.Movie"
                },
                "movieHall": {
                    "$ref": "#/definitions/model.MovieHall"
                },
                "movieHallId": {
                    "type": "integer"
                },
                "movieId": {
                    "type": "integer"
                },
                "start_of_screening": {
                    "type": "string"
                }
            }
        }
    }
}