# movietown/model/screening.go
package model

import (
	"time"
)

type Screening struct {
	ID                 uint `gorm:"primaryKey,foreignKey=screening_id"`
	MovieId            *uint
	Movie              Movie `gorm:"foreignKey:MovieId"`
	MovieHallId        *uint
	MovieHall          MovieHall `gorm:"foreignKey:MovieHallId"`
	Start_of_screening time.Time
}